Summary: 64-bit fixes
Remove the .in data files.
Make System::get_oparg return intptr_t, not int, so we work on 64-bit systems properly.
Change the String operator[] definition a bit so that we work on more G++ versions.
Fix fileobj.cc in read_token to use ints rather than chars, so we don't crash on OpenPOWER.

+++ include/scriptix/system.h
@@ -102,7 +102,7 @@ class SScriptManager : public IManager {
 	inline Value get_value (void) { return data_stack.back(); }
 
 	// Fetch item from frame stack for op atrgs; "eats" arg
-	inline int get_oparg (void) { return get_frame().func->nodes[get_frame().op_ptr++]; }
+	inline intptr_t get_oparg (void) { return get_frame().func->nodes[get_frame().op_ptr++]; }
 
 	// Manipulate data stack - INLINE for speed
 	inline int
